Major internal refactoring
The main goal of this commit, besides small stuff, is removing the dependency of
frameworks. You should never do an import AVFoundation outside the relevant classes
that implement your AudioPlayerService (or something like that). That's my philosophy
anyways. This means you could very easily change the audio framework. That's the
state I'm trying to get to.