Use Repository pattern between business logic and database
This is the first draft of the repository pattern between the business
logic and the database layer. Now there is some duplication, but it's
still better than having one class that does it all. I wonder if it
could be simplified still, but right now it's very easy to use, so I'll
see it as an improvement.