Turn three repositories into one generic RealmRepository<T>
Gone is the duplication, now there is only one. With the power of Swift
generics I can have one RealmRepository with code that looks like there
is only one type that can be handled, only there is as many Object
subclasses as I want. Swift infers at the call site, so I don't have to
specify the generic argument, only something like

    let project = Project()
    RealmRepository().save(object: project)

Only when I update an object do I need to specify the generic argument.

    let project = Project()
    RealmRepository<Project>().update(id: project.id, value: "Hello",
    keyPath: \.title)

The reason is that we pass in the ID of the object, not the object
itself. This because the parameter is a constant, so it cannot be
changed. For now this is fine, but might consider using an inout
parameter or just passing in the object but still fetching it from Realm
and updating that copy and then updating it. It's not really a nice
solution, but it could work.