Remove concept of section
I have been thinking for a long time that maybe the concept of a section
was redundant, or at least too heavy for this app. I'm now thinking that
a mere tagging system could be better. This is also more in line with
the new philosophy of the app that you want to have the recording front
and center, and not all of this CRUDing. Then you can later tag the
recordings with a section, so that a section is nothing more than a text
you slap onto the recording. Much better, if you ask me.