Hack around Core Data child view deletion issue
The issue was that if I deleted a project in the project view, the app
would crash with an inconsistency exception, specifically saying that
the number of rows in the recordings list was invalid before vs after
the deletion. This was solved by posting a notification and waiting a
bit with the deletion. It's a hack, it's not pretty and I hope there is
a better way.